requirements.txt
streamlit
openai
pdfplumber
python-docx
pymongo[srv]
pandas
openpyxl
python-dotenv
dnspython
